Restructured API Routes for maintainability.
